+++ R/addGOToGraph.R
@@ -20,7 +20,6 @@
 #' 
 #' @examples 
 #' ## This function is internal
-#' attach(environment(FELLA:::addCellularComponentToGraph))
 #' library(igraph)
 #' data(FELLA.sample)
 #' data(input.sample)
@@ -33,8 +32,9 @@
 #' threshold = 0.1, 
 #' object = obj, 
 #' data = FELLA.sample)
+#' g
 #' ## Add the cellular component
-#' g.cc <- addGOToGraph(
+#' g.cc <- FELLA:::addGOToGraph(
 #' graph = g, 
 #' GOterm = "GO:0005739")
 #'