+++ R/plotBipartite.R
@@ -16,14 +16,13 @@
 #' 
 #' @examples 
 #' ## This function is internal
-#' attach(environment(FELLA:::plotBipartite))
-#' 
 #' data(FELLA.sample)
 #' data(input.sample)
 #' ## Enrich input
 #' obj <- enrich(
 #' compounds = input.sample, 
-#' data = FELLA.sample)
+#' data = FELLA.sample, 
+#' method = "hypergeom")
 #' ## Generate the bipartite graph (only in the hypergeometric test)
 #' g <- generateResultsGraph(
 #' method = "hypergeom", 
@@ -31,7 +30,7 @@
 #' object = obj, 
 #' data = FELLA.sample)
 #' ## Plot it
-#' plotBipartite(g)
+#' FELLA:::plotBipartite(g)
 #' 
 #' @import igraph
 plotBipartite <- function(